# grevm(opt): optimize cache state