- Apply evm state changes to state cache without generating transition.
- Directly merge partition state cache to scheduler if there are no conflicts in the entire partition.

Before opt.
```
Independent Raw Transfers/Grevm Parallel
                        time:   [71.706 ms 73.537 ms 77.904 ms]
Independent Raw Transfers/Grevm Sequential
                        time:   [115.13 ms 117.05 ms 118.41 ms]
Dependent Raw Transfers/Grevm Parallel
                        time:   [95.392 ms 96.752 ms 98.168 ms]
Dependent Raw Transfers/Grevm Sequential
                        time:   [127.72 ms 128.53 ms 129.83 ms]
Independent ERC20/Grevm Parallel
                        time:   [94.768 ms 96.723 ms 99.558 ms]
Independent ERC20/Grevm Sequential
                        time:   [198.46 ms 200.44 ms 202.76 ms]
Dependent ERC20/Grevm Parallel
                        time:   [128.26 ms 133.78 ms 137.61 ms]
Dependent ERC20/Grevm Sequential
                        time:   [220.80 ms 223.12 ms 225.42 ms]
```
After opt.
```
Independent Raw Transfers/Grevm Parallel
                        time:   [65.826 ms 66.189 ms 66.474 ms]
Independent Raw Transfers/Grevm Sequential
                        time:   [99.834 ms 100.71 ms 101.30 ms]
Dependent Raw Transfers/Grevm Parallel
                        time:   [82.873 ms 84.296 ms 85.306 ms]
Dependent Raw Transfers/Grevm Sequential
                        time:   [108.55 ms 111.08 ms 112.60 ms]
Independent ERC20/Grevm Parallel
                        time:   [81.495 ms 82.734 ms 84.965 ms]
Independent ERC20/Grevm Sequential
                        time:   [186.97 ms 189.85 ms 193.19 ms]
Dependent ERC20/Grevm Parallel
                        time:   [113.41 ms 114.57 ms 116.73 ms]
Dependent ERC20/Grevm Sequential
                        time:   [196.04 ms 196.59 ms 197.16 ms]
```